# openapispec.yml
openapi: 3.0.0
info:
  title: Shopping List API
  version: 1.0.1
servers:
  - url: http://localhost:3000
    description: Local server
paths:
...
  /api/items/{id}:
    get:
      summary: Artikel nach ID abrufen
      operationId: getItemById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Details des Artikels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Artikel nicht gefunden
    put:
      summary: Artikel aktualisieren
      operationId: updateItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '200':
          description: Artikel erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Ungültige Anfrage
        '404':
          description: Artikel nicht gefunden
    delete:
      summary: Artikel löschen
      operationId: deleteItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Artikel erfolgreich gelöscht
        '404':
          description: Artikel nicht gefunden
...
components:
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
...